server/Wav2Lip/requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
uuid
# Add SadTalker dependencies separately per their repo
